feat(core): Story 2.2 — extract LLM resilience to llm_client.py, zero behavior change
Extract the retry/fallback/circuit-breaker logic out of
backend/pipeline/narrative_generator.py into a reusable
backend/core/llm_client.py so the Reporting Agent (2.3) and Monitoring
Agent (2.4) can invoke LLM calls without duplicating resilience machinery.

- New backend/core/ package (__init__.py + llm_client.py).
- LLMClient owns the provider loop, _call_claude/_call_openai/_call_gemini,
  resilience constants, and the system prompt; generate() is renamed
  generate_narrative(payload_json, pipeline_run_id).
- NarrativeGenerator becomes a thin Stage-3 delegator (287 -> ~37 lines);
  public generate(payload, run_id) surface unchanged.
- Provider SDK imports remain lazy inside their _call_* methods; no new
  dependency added to pyproject.toml.
- structlog stage binding and all event names preserved byte-for-byte.
- Tests repointed to backend.core.llm_client.LLMClient targets; assertions
  unchanged. 111 passed / 1 skipped (anthropic absent), zero regressions.
