+++ backend/core/__init__.py
@@ -1,0 +1,1 @@
+# backend/core — shared utilities (LLM client, future: cost tracking)
